add stub add_animation method to GenericXEnemy class
Since _animgr is private filed, and there are not any other ways to
add animation beside AnimationManager method, we need to add a
function to AnimationManager's owner class.

We call this function add_animation and it just passes *args
and **kwargs to the add method of _animgr.

in scope of #50