remove raise from AnimationManager animation stop
We raised StopIteration exception from objects of Animation class,
when animation finished inside the AnimationManager instance. It was
ok, but quite inconvenient to handle. Anyway, animation change
mechanism is not implemented yet.

So for now just replace that rise with plain return, and don't support
AnimationManager objects with more than one animation.

We also change test_animation accordingly.

in scope of #50