implement Level support for xoinvader game
Well, actually it's just renaming EnemyWave to Level, since EnemyWave
turned out to be just that type of sequence handler that makes it fit
for xoinvader level support.

Also this commit moves all game objects in TestLevel class in InGame
state, as they all are part of the level, not the state. This division
came quite natural, which is good. Now we can proceed to move test
level to separate module and to create something more elaborate, with
more enemies and backgrounds, and expiriment with level utils, like
enemy ships cleanup and so on.

Tests and imports were altered according to name changes. No logic
changes in the engine were implemented.

issue #22