collision: make TypePair not commutative
I want to change collision registering scheme:
Handler registered by decorator to some mapping TypePair to handler
list, but when handler fired I want to call handler as method of
object, handler connected to and pass second object as argument.

To do calls to right object we need to keep order, otherwise we will
call all handlers with semi-random order, what is not great.

Assume we have:
* GameObjects obj1 and obj2
* Colliders col1 and col2
* Collision handlers handler1 and handler2

Col1 is attached to obj1, col2 is attached to obj2.
Handler1 used for "col1_col2" pair, handler2 used for
"col2_col1" pair.

When handler1 is fired I want to do in like this:
handler1(collider1.obj, collider2.obj)

To fire handler in right order we need ordering in TypePair.

Signed-off-by: Pavel Kulyov <kulyov.pavel@gmail.com>