collision: make Collider work with new CollisionManager
Now colliders.
From programmer's sight, handlers tight together two collidable
GameObjects, not thier colliders. So we dont need Collider.add_handler
method. But the greatest reason to delete this METHOD is because this
is METHOD. So we need OBJECT to be created at IMPORT time. So it might
be for example CLASSFIELD. But this is impossible because Collider
must track object's position. Position is part of state, part of
Object's INSTANCE.

Summary:
* adding handlers is IMPORTTIME.
* adding colliders is RUNTIME.

Signed-off-by: Pavel Kulyov <kulyov.pavel@gmail.com>