application,ingame,ship,state: add scoring, add postinit stage
On State.__init__ stage may be objects that want to get access to
current State. But when State is initializing object is not ready.
Adding State.postinit stage workarounds this behavior. Put objects
that want to use State methods in State.postinit and all will be
great.

Current case is:
 State.__init__->PlayerShip.__init__->Collider.__init__->
 ->application->state.

Last step of this chain causes exception because no such state
in application as interpreter stands on State.__init__ at high
execution frame.

Signed-off-by: Pavel Kulyov <kulyov.pavel@gmail.com>